refactor(web): profile avatar shared ui로 분리
