+++ apps/web/src/shared/ui/index.ts
@@ -1,3 +1,4 @@
 export { AuthCard } from "./auth-card";
 export { DatePicker } from "./date-picker";
+export { ProfileAvatar } from "./profile-avatar";
 export { TimePicker } from "./time-picker";
