+++ apps/web/src/shared/ui/profile-icon.tsx
